routes protection, CRUD only by loggedin user done

+++ controllers/contactController.js
@@ -3,9 +3,9 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all the contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler( async (req, res) => {
-    const contacts = await Contact.find(); // added this after making the model, and we will get all the contacts from the database, and we will send it to the client.
+    const contacts = await Contact.find({user_id: req.user.id}); //?the uder_id: req.user.id is for validating the user // added this after making the model, and we will get all the contacts from the database, and we will send it to the client.
     // res.status(200).json({ message: "Get all contacts" }); //! when we haven't used the database, we used this line, but now we will use the database, so we will comment this line.
     res.status(200).json(contacts);
 });
@@ -13,7 +13,7 @@ const getContacts = asyncHandler( async (req, res) => {
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler( async (req, res) => {
     console.log("The request body is: ", req.body);
     const { name, email, phone, type } = req.body;
@@ -25,6 +25,7 @@ const createContact = asyncHandler( async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     // res.status(201).json({ message: "Create contact" }); //! this was before using the database, now we will use the database, so we will comment this line.
 
@@ -34,7 +35,7 @@ const createContact = asyncHandler( async (req, res) => {
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler( async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -45,12 +46,17 @@ const getContact = asyncHandler( async (req, res) => {
 });
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler( async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not authorized");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -66,7 +72,7 @@ const updateContact = asyncHandler( async (req, res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler( async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -74,10 +80,15 @@ const deleteContact = asyncHandler( async (req, res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not authorized");
+    }
+
     await contact.deleteOne({ _id: req.params.id });
 
 
-    res.status(200).json({message : "deleted"});
+    res.status(200).json(contact);
 });
 
 module.exports = { getContacts, createContact, getContact, updateContact, deleteContact };
